backend(dev): fix dev startup and domain types
- Remove duplicate @EnableJpaAuditing (AuditingConfig handles it)
- Disable Flyway in dev; rely on JPA create-drop + data-dev.sql
- Standardize temperature fields to BigDecimal in entities/services/alerts
- Update services and alert evaluator to handle BigDecimal
- Leave default temperature null for new machines
- Misc config cleanups

+++ src/main/java/com/example/coffeemachine/alert/AlertEvaluatorService.java
@@ -101,7 +101,7 @@ public class AlertEvaluatorService {
             return;
         }
         
-        double temp = machine.getTemperature();
+        double temp = machine.getTemperature().doubleValue();
         
         if (temp < MIN_TEMPERATURE || temp > MAX_TEMPERATURE) {
             String message = temp < MIN_TEMPERATURE ?
